@@ -12,13 +12,13 @@ class BottomSheetSelectResult {
   BottomSheetSelectResult({required this.cancelled});
 }
 
-class BottomSheetSelect extends StatefulWidget {
-  final String selectedId;
+class BottomSheetSelect<T> extends StatefulWidget {
+  final T selectedId;
   final String title;
   final Widget? description;
-  final List<Option> items;
+  final List<Option<T>> items;
   final bool showSelection;
-  final void Function(String id) onSelectionChanged;
+  final void Function(T id) onSelectionChanged;
   final bool popOnChange;
 
   const BottomSheetSelect({
@@ -33,11 +33,11 @@ class BottomSheetSelect extends StatefulWidget {
   }) : super(key: key);
 
   @override
-  State<BottomSheetSelect> createState() => _BottomSheetSelectState();
+  State<BottomSheetSelect<T>> createState() => _BottomSheetSelectState<T>();
 }
 
-class _BottomSheetSelectState extends State<BottomSheetSelect> {
-  late String localSelectedId;
+class _BottomSheetSelectState<T> extends State<BottomSheetSelect<T>> {
+  late T localSelectedId;
   @override
   void initState() {
     super.initState();
@@ -49,15 +49,15 @@ class _BottomSheetSelectState extends State<BottomSheetSelect> {
     final design = DesignSystem.of(context);
     return Container(
       color: design.colors.background1,
-      child: SafeArea(
-        child: Padding(
-          padding: screenInsets(context) + const EdgeInsets.only(left: 16, right: 16, top: 8, bottom: 32),
+      child: Padding(
+        padding: screenInsets(context) + const EdgeInsets.only(left: 16, right: 16, top: 8, bottom: 32),
+        child: SafeArea(
           child: Column(
-            mainAxisAlignment: MainAxisAlignment.center,
             mainAxisSize: MainAxisSize.min,
             children: [
               Container(
-                margin: const EdgeInsets.only(bottom: 16),
+                alignment: Alignment.center,
+                margin: const EdgeInsets.only(bottom: 8),
                 child: Container(
                   height: 5,
                   width: 40,
@@ -67,48 +67,63 @@ class _BottomSheetSelectState extends State<BottomSheetSelect> {
                   ),
                 ),
               ),
-              Container(
-                margin: const EdgeInsets.only(bottom: 16),
-                child: Text(
-                  widget.title,
-                  textAlign: TextAlign.center,
-                  style: design.textStyles.title3,
-                ),
-              ),
-              if (widget.description != null) widget.description!,
               Flexible(
-                child: OptionList(
-                  optionData: widget.items,
-                  currentSelection: localSelectedId,
-                  showSelection: widget.showSelection,
-                  onSelectionChange: (val) {
-                    if (val == null) return;
-                    setState(() {
-                      localSelectedId = val;
-                    });
-                    widget.onSelectionChanged(val);
-                    if (widget.popOnChange) {
-                      Navigator.pop(context, BottomSheetSelectResult(cancelled: false));
-                    }
-                  },
+                child: CustomScrollView(
+                  primary: false,
+                  shrinkWrap: true,
+                  scrollBehavior: const ScrollBehavior().copyWith(overscroll: false),
+                  physics: const ClampingScrollPhysics(),
+                  slivers: [
+                    SliverToBoxAdapter(
+                      child: Container(
+                        margin: const EdgeInsets.only(top: 16, bottom: 24),
+                        child: Text(
+                          widget.title,
+                          textAlign: TextAlign.center,
+                          style: design.textStyles.title3,
+                        ),
+                      ),
+                    ),
+                    if (widget.description != null) SliverToBoxAdapter(child: widget.description!),
+                    SliverToBoxAdapter(
+                      child: OptionList<T>(
+                        optionData: widget.items as List<Option<T>>,
+                        currentSelection: localSelectedId,
+                        showSelection: widget.showSelection,
+                        onSelectionChange: (val) {
+                          if (val == null) return;
+                          setState(() {
+                            localSelectedId = val;
+                          });
+                          widget.onSelectionChanged(val);
+                          if (widget.popOnChange) {
+                            Navigator.pop(context, BottomSheetSelectResult(cancelled: false));
+                          }
+                        },
+                      ),
+                    ),
+                  ],
                 ),
               ),
-              Container(
-                margin: const EdgeInsets.only(top: 16),
-                width: double.infinity,
-                height: 52.1,
-                child: TextButton(
-                  style: TextButton.styleFrom(
-                    backgroundColor: design.colors.separatorOnLight,
-                    shape: const RoundedRectangleBorder(borderRadius: BorderRadius.all(Radius.circular(100))),
-                    side: BorderSide(width: 1, color: design.colors.separatorOnLight),
-                  ),
-                  onPressed: () {
-                    Navigator.pop(context, BottomSheetSelectResult(cancelled: true));
-                  },
-                  child: Text(
-                    S.of(context).cancel,
-                    style: design.textStyles.button1.copyWith(color: design.colors.label1),
+              Align(
+                alignment: Alignment.bottomCenter,
+                child: Container(
+                  margin: const EdgeInsets.only(top: 16),
+                  width: double.infinity,
+                  height: 52.1,
+                  child: TextButton(
+                    style: TextButton.styleFrom(
+                      backgroundColor: design.colors.separatorOnLight,
+                      shape: const RoundedRectangleBorder(borderRadius: BorderRadius.all(Radius.circular(100))),
+                      side: BorderSide(width: 1, color: design.colors.separatorOnLight),
+                    ),
+                    onPressed: () {
+                      Navigator.pop(context, BottomSheetSelectResult(cancelled: true));
+                    },
+                    child: Text(
+                      S.of(context).cancel,
+                      style: design.textStyles.button1.copyWith(color: design.colors.label1),
+                    ),
                   ),
                 ),
               ),
