@@ -143,39 +143,42 @@ class _AppRootState extends ConsumerState<AppRoot> {
           ),
           child: GraphQLProvider(
             client: ValueNotifier(ref.watch(gqlClientProvider)),
-            child: MaterialApp.router(
-              localizationsDelegates: S.localizationsDelegates,
-              localeResolutionCallback: (locale, supportedLocales) {
-                if (locale?.languageCode == 'no') {
-                  return const Locale('nb');
-                }
-                if (supportedLocales.map((e) => e.languageCode).contains(locale?.languageCode)) {
-                  return locale;
-                }
-                return Locale(FlavorConfig.current.defaultLanguage);
-              },
-              supportedLocales: S.supportedLocales,
-              locale: ref.watch(settingsProvider).appLanguage,
-              theme: ThemeData(),
-              darkTheme: DesignSystem.of(context).materialThemeData,
-              themeMode: ThemeMode.dark,
-              debugShowCheckedModeBanner: false,
-              title: 'BCC Media',
-              routerDelegate: widget.appRouter.delegate(
-                initialRoutes: [const AutoLoginScreenRoute()],
-                navigatorObservers: () => [AnalyticsNavigatorObserver()],
+            child: Directionality(
+              textDirection: TextDirection.ltr,
+              child: MaterialApp.router(
+                localizationsDelegates: S.localizationsDelegates,
+                localeResolutionCallback: (locale, supportedLocales) {
+                  if (locale?.languageCode == 'no') {
+                    return const Locale('nb');
+                  }
+                  if (supportedLocales.map((e) => e.languageCode).contains(locale?.languageCode)) {
+                    return locale;
+                  }
+                  return Locale(FlavorConfig.current.defaultLanguage);
+                },
+                supportedLocales: S.supportedLocales,
+                locale: ref.watch(settingsProvider).appLanguage,
+                theme: ThemeData(),
+                darkTheme: DesignSystem.of(context).materialThemeData,
+                themeMode: ThemeMode.dark,
+                debugShowCheckedModeBanner: false,
+                title: 'BCC Media',
+                routerDelegate: widget.appRouter.delegate(
+                  initialRoutes: [const AutoLoginScreenRoute()],
+                  navigatorObservers: () => [AnalyticsNavigatorObserver()],
+                ),
+                routeInformationParser: widget.appRouter.defaultRouteParser(includePrefixMatches: true),
+                builder: (BuildContext context, Widget? child) {
+                  return ResponsiveBreakpoints.builder(
+                    child: MediaQuery(
+                      data: MediaQuery.of(context).copyWith(textScaleFactor: 1.0),
+                      child: child!,
+                    ),
+                    breakpoints: breakpoints,
+                    breakpointsLandscape: breakpoints,
+                  );
+                },
               ),
-              routeInformationParser: widget.appRouter.defaultRouteParser(includePrefixMatches: true),
-              builder: (BuildContext context, Widget? child) {
-                return ResponsiveBreakpoints.builder(
-                  child: MediaQuery(
-                    data: MediaQuery.of(context).copyWith(textScaleFactor: 1.0),
-                    child: child!,
-                  ),
-                  breakpoints: breakpoints,
-                  breakpointsLandscape: breakpoints,
-                );
-              },
             ),
           ),
         ),
