@@ -9,59 +9,55 @@ import 'package:hooks_riverpod/hooks_riverpod.dart';
 import '../../theme/design_system/design_system.dart';
 
 class Avatar extends HookConsumerWidget {
-  const Avatar({Key? key}) : super(key: key);
+  const Avatar({
+    Key? key,
+    this.width = 100.0,
+    this.backgroundColor,
+  }) : super(key: key);
+
+  final double? width;
+  final Color? backgroundColor;
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
-    final imageUrl = ref.read(authStateProvider).user?.picture.toString();
-    final name = ref.read(authStateProvider).user?.name;
-    const avatarWidth = 100.0;
-
-    return Container(
-      margin: const EdgeInsets.only(top: 24, bottom: 16),
-      child: Column(
-        children: [
-          Container(
-            margin: const EdgeInsets.only(bottom: 16),
-            child: kIsWeb
-                ? Container(
-                    alignment: Alignment.center,
-                    width: avatarWidth,
-                    height: avatarWidth,
-                    child: SvgPicture.string(SvgIcons.avatar),
-                  )
-                : Container(
-                    width: avatarWidth,
-                    height: avatarWidth,
-                    decoration: BoxDecoration(
-                      shape: BoxShape.circle,
-                      border: Border.all(
-                        color: DesignSystem.of(context).colors.background2,
-                        width: 2,
+    final imageUrl = ref.watch(authStateProvider.select((value) => value.user?.picture));
+    final design = DesignSystem.of(context);
+    return kIsWeb
+        ? Container(
+            alignment: Alignment.center,
+            width: width,
+            height: width,
+            child: SvgPicture.string(SvgIcons.avatar),
+          )
+        : Container(
+            width: width,
+            height: width,
+            decoration: BoxDecoration(
+              color: backgroundColor ?? design.colors.background2,
+              shape: BoxShape.circle,
+              border: Border.all(
+                color: backgroundColor ?? design.colors.background2,
+                width: 2,
+              ),
+            ),
+            child: ClipRRect(
+              borderRadius: BorderRadius.circular(1000),
+              child: Image(
+                image: NetworkImage(imageUrl ?? ''),
+                errorBuilder: (BuildContext context, Object exception, StackTrace? stackTrace) {
+                  return FractionallySizedBox(
+                    widthFactor: 0.4,
+                    child: Center(
+                      child: SvgPicture.string(
+                        SvgIcons.avatar,
+                        colorFilter: ColorFilter.mode(DesignSystem.of(context).colors.tint1, BlendMode.srcIn),
+                        height: 35,
                       ),
                     ),
-                    child: ClipRRect(
-                      borderRadius: BorderRadius.circular(1000),
-                      child: Image(
-                        image: NetworkImage(imageUrl ?? ''),
-                        errorBuilder: (BuildContext context, Object exception, StackTrace? stackTrace) {
-                          return Center(
-                            child: SvgPicture.string(
-                              SvgIcons.avatar,
-                              color: DesignSystem.of(context).colors.tint1,
-                            ),
-                          );
-                        },
-                      ),
-                    ),
-                  ),
-          ),
-          Text(
-            name ?? '',
-            style: DesignSystem.of(context).textStyles.title1,
-          ),
-        ],
-      ),
-    );
+                  );
+                },
+              ),
+            ),
+          );
   }
 }
