@@ -1,8 +1,11 @@
+import 'dart:math';
+
 import 'package:flutter/material.dart';
 
 import '../../helpers/time.dart';
 import '../../helpers/episode_state.dart';
 import '../../models/episode_thumbnail_data.dart';
+import '../badges/download_expires_badge.dart';
 import 'misc/bordered_image_container.dart';
 import 'misc/episode_duration.dart';
 import 'misc/watch_progress_indicator.dart';
@@ -14,6 +17,8 @@ class EpisodeThumbnail extends StatelessWidget {
   final double? imageHeight;
   final double aspectRatio;
   final bool isLive;
+  final bool? useCache;
+  final DateTime? expiresAt;
 
   const EpisodeThumbnail({
     super.key,
@@ -22,18 +27,24 @@ class EpisodeThumbnail extends StatelessWidget {
     this.imageHeight,
     this.aspectRatio = 16 / 9,
     this.isLive = false,
+    this.useCache,
+    this.expiresAt,
   });
 
   factory EpisodeThumbnail.withSize({
     required EpisodeThumbnailData episode,
     required Size imageSize,
     bool isLive = false,
+    bool? useCache,
+    DateTime? expiresAt,
   }) {
     return EpisodeThumbnail(
       episode: episode,
       imageWidth: imageSize.width,
       imageHeight: imageSize.height,
       isLive: isLive,
+      useCache: useCache,
+      expiresAt: expiresAt,
     );
   }
 
@@ -42,12 +53,16 @@ class EpisodeThumbnail extends StatelessWidget {
     required double imageWidth,
     required double aspectRatio,
     bool isLive = false,
+    bool? useCache,
+    DateTime? expiresAt,
   }) {
     return EpisodeThumbnail(
       imageWidth: imageWidth,
       aspectRatio: aspectRatio,
       isLive: isLive,
       episode: episode,
+      useCache: useCache,
+      expiresAt: expiresAt,
     );
   }
 
@@ -60,18 +75,35 @@ class EpisodeThumbnail extends StatelessWidget {
       child: imageHeight == null
           ? AspectRatio(
               aspectRatio: aspectRatio,
-              child: EpisodeThumbnailStack(episode: episode, isLive: isLive),
+              child: _EpisodeThumbnailStack(
+                episode: episode,
+                isLive: isLive,
+                useCache: useCache,
+                expiresAt: expiresAt,
+              ),
             )
-          : EpisodeThumbnailStack(episode: episode, isLive: isLive),
+          : _EpisodeThumbnailStack(
+              episode: episode,
+              isLive: isLive,
+              useCache: useCache,
+              expiresAt: expiresAt,
+            ),
     );
   }
 }
 
-class EpisodeThumbnailStack extends StatelessWidget {
+class _EpisodeThumbnailStack extends StatelessWidget {
   final EpisodeThumbnailData episode;
   final bool isLive;
+  final bool? useCache;
+  final DateTime? expiresAt;
 
-  const EpisodeThumbnailStack({required this.episode, required this.isLive, super.key});
+  const _EpisodeThumbnailStack({
+    required this.episode,
+    required this.isLive,
+    required this.useCache,
+    required this.expiresAt,
+  });
 
   bool get watched => episode.progress != null && episode.progress! > episode.duration * 0.9;
 
@@ -90,9 +122,9 @@ class EpisodeThumbnailStack extends StatelessWidget {
         episode.locked && !isLive
             ? Opacity(
                 opacity: 0.5,
-                child: BorderedImageContainer(imageUrl: episode.image),
+                child: BorderedImageContainer(imageUrl: episode.image, useCache: useCache),
               )
-            : BorderedImageContainer(imageUrl: episode.image),
+            : BorderedImageContainer(imageUrl: episode.image, useCache: useCache),
         if (episode.locked && !isLive)
           Container(
             width: double.infinity,
@@ -120,6 +152,7 @@ class EpisodeThumbnailStack extends StatelessWidget {
               child: Row(
                 children: [
                   if (watched) const WatchedBadge(),
+                  if (expiresAt != null) DownloadExpiresBadge(expiresAt: expiresAt!),
                   const Spacer(),
                   EpisodeDuration(duration: getFormattedDuration(episode.duration)),
                 ],
