@@ -1,5 +1,7 @@
 import 'package:brunstadtv_app/helpers/transparent_image.dart';
+import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_cache_manager/flutter_cache_manager.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:flutter_image/network.dart';
 
@@ -70,15 +72,16 @@ Widget simpleFadeInImage({required String url, Duration? duration}) => LayoutBui
 ImageProvider<Object> networkImageWithRetryAndResize({
   required String imageUrl,
   int? cacheHeight,
-}) =>
-    ResizeImage.resizeIfNeeded(
-      null,
-      cacheHeight,
-      NetworkImageWithRetry(
-        imageUrl,
-        headers: const {'Keep-Alive': 'timeout=20, max=5'},
-      ),
-    );
+}) {
+  return ResizeImage.resizeIfNeeded(
+    null,
+    cacheHeight,
+    NetworkImageWithRetry(
+      imageUrl,
+      headers: const {'Keep-Alive': 'timeout=20, max=5'},
+    ),
+  );
+}
 
 Widget imageErrorBuilder(
   BuildContext context,
