@@ -55,7 +55,7 @@ class Auth0IdToken {
 
   bool? get mediaSubscriber => userMetadata?['media_subscriber'];
   bool? get hasMembership => appMetadata?['hasMembership'];
-  String? get bccPersonId => appMetadata?['personId'].toString();
+  String? get bccPersonId => appMetadata?['personId']?.toString();
 
   @JsonKey(name: 'auth_time')
   final int? authTime; // this might be null for the first time login
